+++ src/main.js
@@ -3,13 +3,13 @@ import { addRoutes, onNavigate } from './router/index.js';
 import {Home} from './components/Home';
 import {Login} from './components/Login';
 import {Register} from './components/Register';
-import { loginGoogle } from './lib/autenticar.js';
+import { Feed } from './components/Feed';
 
 addRoutes({
   '/': Home,
   '/login': Login,
   '/register': Register,
-  '/loginGoogle': loginGoogle,
+  '/feed': Feed,
 });
 // Lógica de la aplicacion
 window.onload = () => {
